Abstracting normalizeRemoteLoaderSlug helper
